+++ be/src/storage/lake/spark_load.cpp
@@ -2,11 +2,9 @@
 
 #include "storage/lake/spark_load.h"
 
-#include "runtime/runtime_state.h"
-#include "storage/chunk_helper.h"
 #include "storage/lake/filenames.h"
 #include "storage/lake/tablet_writer.h"
-#include "storage/push_handler.h"
+#include "util/push_broker_reader.h"
 #include "storage/storage_engine.h"
 #include "storage/tablet_manager.h"
